LoginServer/requirements.txt
fastapi~=0.95.1
starlette~=0.26.1
pydantic~=1.10.7
PyJWT~=2.7.0
passlib~=1.7.4
pymongo~=4.3.3
odmantic~=0.9.2